segregation of concerns
